Move offscreen renderer to base viewer dir.
