+++ src/mjlab/viewer/__init__.py
@@ -5,5 +5,6 @@ from mjlab.viewer.base import EnvProtocol as EnvProtocol
 from mjlab.viewer.base import PolicyProtocol as PolicyProtocol
 from mjlab.viewer.base import VerbosityLevel as VerbosityLevel
 from mjlab.viewer.native import NativeMujocoViewer as NativeMujocoViewer
+from mjlab.viewer.offscreen_renderer import OffscreenRenderer as OffscreenRenderer
 from mjlab.viewer.viewer_config import ViewerConfig as ViewerConfig
 from mjlab.viewer.viser import ViserPlayViewer as ViserPlayViewer
