+++ src/mjlab/viewer/native/__init__.py
@@ -1,6 +1,5 @@
 """Native MuJoCo viewer implementation."""
 
-from mjlab.viewer.native.renderer import OffscreenRenderer as OffscreenRenderer
 from mjlab.viewer.native.viewer import NativeMujocoViewer as NativeMujocoViewer
 from mjlab.viewer.native.visualizer import (
   MujocoNativeDebugVisualizer as MujocoNativeDebugVisualizer,
